requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-generativeai==0.3.1